Working on styles and renderings
